requirements.txt
africastalking==1.2.4
alembic==1.6.5
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
configparser==5.0.2
contextlib2==21.6.0
Flask==2.0.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
Pillow==8.3.0
pkg-resources==0.0.0
protobuf==3.17.3
psycopg2-binary==2.9.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
requests==2.25.1
schema==0.7.4
six==1.16.0
SQLAlchemy==1.4.20
urllib3==1.26.6
Werkzeug==2.0.1
